reorganize top-level API into submodule folders
