+++ compose/index.js
