+++ delegate/index.js
